+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineTest.kt
@@ -399,4 +399,16 @@ class GeckoEngineTest {
         assertTrue(onErrorCalled)
         assertSame(exception, throwable)
     }
+
+    @Test
+    fun `test parsing engine version`() {
+        val runtime: GeckoRuntime = mock()
+        val engine = GeckoEngine(context, runtime = runtime)
+        val version = engine.version
+
+        println(version)
+
+        assertTrue(version.major >= 69)
+        assertTrue(version.isAtLeast(69, 0, 0))
+    }
 }